Catalog update task fixes (#141)
