+++ shopelectro/tasks.py
@@ -1,6 +1,9 @@
+from datetime import timedelta
+
 from django.core.management import call_command
 
 from shopelectro.celery import app
+from shopelectro.management.commands._update_catalog import utils
 
 
 @app.task
@@ -30,13 +33,17 @@ def update_meta_tags():
     call_command('update_meta_tags')
 
 
-@app.task(autoretry_for=(Exception,), max_retries=3, default_retry_delay=30)
+@app.task(autoretry_for=(Exception,), max_retries=3, default_retry_delay=timedelta(minutes=10))
 def update_catalog():
     # http://docs.celeryproject.org/en/latest/userguide/canvas.html#map-starmap
-    return [
-        update_catalog_command(),
-        update_meta_tags(),
-        generate_price_files(),
-        generate_excel_file(),
-        collect_static()
-    ]
+    try:
+        return [
+            update_catalog_command(),
+            update_meta_tags(),
+            generate_price_files(),
+            generate_excel_file(),
+            collect_static()
+        ]
+    except Exception as exc:
+        utils.report(exc)
+        raise exc
